requirements.txt
tqdm
torch==2.2.1
datasets-2.18.0
transformers==4.39.3